Implement OOP for the classes
