+++ data/duke.txt
@@ -1,3 +1,3 @@
- D |   | hi | by: Apr 05 2013 00:00
- D |   | hi | by: Dec 02 2023 12:34
- E |   | hi | from: Apr 05 2013 12:34 to: Mar 05 2013 12:56
+ D | 0 | hi | by: Apr 05 2013 00:00
+ D | 0 | hi | by: Dec 02 2023 12:34
+ E | 0 | hi | from: Apr 05 2013 12:34 to: Mar 05 2013 12:56
