+++ src/main/java/Dino.java
@@ -1,21 +1,9 @@
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-
-import java.util.ArrayList;
-import java.util.Locale;
 import java.util.Scanner;
 
-import java.io.IOException;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.PrintWriter;
-
 public class Dino {
-    private ArrayList<Task> taskList;
+    private Ui ui;
+    private Storage storage;
+    private TaskList tasks;
 
     public enum TaskType {
         TODO,
@@ -23,339 +11,34 @@ public class Dino {
         EVENT
     }
 
-    TaskType taskType;
-    Task task;
     private static final String FILE_PATH = "./data/duke.txt";
 
-    public Dino() {
-        this.taskList = new ArrayList<>();
-        loadTasksFromFile();
+    public Dino(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+        tasks = storage.loadTasksFromFile();
+
     }
 
-    public static void welcome() {
-        System.out.println("Hola! I'm Dino.\nWhat are you doing here?");
-    }
-
-    public void goodbye() {
-        System.out.println("Goodbye! It was nice meeting you.");
-    }
-
-    public void addTask(Task task) {
-        this.taskList.add(task);
-    }
-
-    public void listTask() {
-        try {
-            if (taskList.isEmpty()) {
-                throw new DinoException("The list is empty.");
-            }
-            System.out.println("Here are the tasks that you wanted to do:");
-            for (int i = 0; i < taskList.size(); i++) {
-                Task task = taskList.get(i);
-                int index = i + 1;
-                System.out.println(index + "." + task);
-            }
-        } catch (DinoException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-    public void deleteTask(int taskNum) throws DinoException {
-        if (taskNum < 1 || taskNum > taskList.size()) {
-            throw new DinoException("Invalid task number. Please provide a valid task number to delete.");
-        }
-
-        Task removedTask = taskList.remove(taskNum - 1);
-
-        System.out.println("Noted. I've removed this task:");
-        System.out.println("  " + removedTask);
-        System.out.println("Now you have " + taskList.size() + " tasks in the list.");
-    }
-
-    private LocalDateTime parseStringToTime(String time) {
-        DateTimeFormatter dateOnlyFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ENGLISH);
-        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm", Locale.ENGLISH);
-
-        LocalDateTime deadlineTime;
-        if (time.contains(" ")) {
-            deadlineTime = LocalDateTime.parse(time, dateTimeFormatter);
-        } else {
-            deadlineTime = LocalDateTime.of(LocalDate.parse(time, dateOnlyFormatter), LocalTime.MIDNIGHT);
-        }
-        return deadlineTime;
-    }
-
-    private String parseStringToNum(String time) {
-        time = time.trim();
-        DateTimeFormatter dateOnlyFormatter = DateTimeFormatter.ofPattern("MMM dd yyyy", Locale.ENGLISH);
-        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm", Locale.ENGLISH);
-
-        LocalDateTime deadlineTime;
-        if (time.contains(" ")) {
-            deadlineTime = LocalDateTime.parse(time, dateTimeFormatter);
-        } else {
-            deadlineTime = LocalDateTime.of(LocalDate.parse(time, dateOnlyFormatter), LocalTime.MIDNIGHT);
-        }
-
-        // Format LocalDateTime to string in "yyyy-MM-dd" format
-        DateTimeFormatter resultFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ENGLISH);
-        String formattedDate = deadlineTime.format(resultFormatter);
-
-        // Append time if available
-        if (deadlineTime.toLocalTime() != LocalTime.MIDNIGHT) {
-            formattedDate += " " + deadlineTime.toLocalTime().format(DateTimeFormatter.ofPattern("HHmm"));
-        }
-
-        return formattedDate;
-    }
-
-
-
-    private Task createTaskFromInput(TaskType taskType, String taskDetails) throws DinoException {
-        switch (taskType) {
-            case TODO:
-                return new ToDo(taskDetails);
-
-            case DEADLINE:
-                String[] deadlineParts = taskDetails.split("/by");
-                if (deadlineParts.length != 2) {
-                    throw new DinoException("Invalid input format for deadline. Please use: deadline <deadline name> /by <time>");
-                }
-                String deadlineName = deadlineParts[0].trim();
-                String deadlineTimeString = deadlineParts[1].trim();
-                if (deadlineName.isEmpty() || deadlineTimeString.isEmpty()) {
-                    throw new DinoException("Deadline name and time cannot be empty.");
-                }
-
-                try {
-                    return new Deadline(deadlineName, parseStringToTime(deadlineTimeString));
-                } catch (DateTimeParseException e) {
-                    throw new DinoException("Error parsing deadline date and time: " + e.getMessage());
-                }
-
-            case EVENT:
-                String[] eventParts = taskDetails.split("/from|/to");
-                if (eventParts.length != 3) {
-                    throw new DinoException("Invalid input format for event. Please use: event <event name> /from <time> /to <time>");
-                }
-                String eventName = eventParts[0].trim();
-                String startTimeString = eventParts[1].trim();
-                String endTimeString = eventParts[2].trim();
-                if (eventName.isEmpty() || startTimeString.isEmpty() || endTimeString.isEmpty()) {
-                    throw new DinoException("Event name, start time, and end time cannot be empty.");
-                }
-
-                try {
-                    LocalDateTime startTime = parseStringToTime(startTimeString);
-                    LocalDateTime endTime = parseStringToTime(endTimeString);
-                    return new Event(eventName, startTime, endTime);
-                } catch (DateTimeParseException e) {
-                    throw new DinoException("Error parsing event date and time: " + e.getMessage());
-                }
-
-            default:
-                throw new DinoException("Unknown task type: " + taskType);
-        }
-    }
-
-    private void saveTasksToFile() {
-        try {
-            File directory = new File("data");
-            if (!directory.exists()) {
-                directory.mkdirs();
-            }
-
-            File file = new File(FILE_PATH);
-            if (!file.exists()) {
-                file.createNewFile();
-            }
-
-            try (PrintWriter writer = new PrintWriter(new FileWriter(file))) {
-                for (Task task : taskList) {
-                    writer.println(task.toString());
-                }
-            }
-        } catch (IOException e) {
-            System.out.println("Error saving tasks to file: " + e.getMessage());
-        }
-    }
-
-
-    private void loadTasksFromFile() {
-        try {
-            File file = new File(FILE_PATH);
-            if (!file.exists()) {
-                System.out.println("Data file does not exist. Creating a new file.");
-                return;
-            }
-
-            Scanner fileScanner = new Scanner(file);
-            while (fileScanner.hasNextLine()) {
-                String taskData = fileScanner.nextLine();
-                String[] parts = taskData.split("\\|");
-
-                if (parts.length > 0) {
-                    String taskTypeString = parts[0].trim();
-                    switch (taskTypeString) {
-                        case "T":
-                            task = createTaskFromInput(TaskType.TODO, parts[2].trim());
-                            break;
-                        case "D":
-                            String[] deadlineParts = parts[3].split(" by: ");
-                            String deadlineDetails = parts[2].trim() + " /by " + parseStringToNum(deadlineParts[1]);
-                            task = createTaskFromInput(TaskType.DEADLINE, deadlineDetails);
-                            break;
-                        case "E":
-                            String[] eventParts = parts[3].split("from:|to:");
-                            String eventDetails = parts[2].trim() + " /from " + parseStringToNum(eventParts[1]) +
-                                    " /to " + parseStringToNum(eventParts[2]);
-                            task = createTaskFromInput(TaskType.EVENT, eventDetails);
-                            break;
-                        default:
-                            System.out.println("Unknown task type in file: " + taskTypeString);
-                    }
-                }
-
-                if (task != null) {
-                    taskList.add(task);
-                } else {
-                    System.out.println("Error loading task from file. Skipping invalid task.");
-                }
-            }
-            fileScanner.close();
-        } catch (FileNotFoundException e) {
-            System.out.println("Error loading tasks from file: " + e.getMessage());
-        } catch (DinoException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-
-    private void handleTaskCreation(Scanner sc, TaskType taskType) {
-        try {
-            String taskDetails = sc.nextLine().trim();
-            if (taskDetails.isEmpty()) {
-                throw new DinoException("Description cannot be empty.");
-            }
-
-            addTask(createTaskFromInput(taskType, taskDetails));
-
-            System.out.println("Okay.");
-            System.out.println("  " + taskList.get(taskList.size() - 1));
-            System.out.println("Now you have " + taskList.size() + " in the list.");
-        } catch (DinoException e) {
-            System.out.println("Error: " + e.getMessage());
-        }
-    }
-
-    private void printTasksForDate(Scanner sc) {
-        try {
-            String dateString = sc.next();
-            DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ENGLISH);
-            LocalDate date = LocalDate.parse(dateString, dateFormatter);
-
-            System.out.println("Tasks for " + date + ":");
-
-            taskList.stream()
-                    .filter(task -> task instanceof Deadline)
-                    .map(task -> (Deadline) task)
-                    .filter(deadline -> deadline.getDateTime().toLocalDate().equals(date))
-                    .forEach(System.out::println);
-
-            taskList.stream()
-                    .filter(task -> task instanceof Event)
-                    .map(task -> (Event) task)
-                    .filter(event -> event.getStartTime().toLocalDate().equals(date) || event.getEndTime().toLocalDate().equals(date))
-                    .forEach(System.out::println);
-
-        } catch (DateTimeParseException e) {
-            System.out.println("Error parsing date: " + e.getMessage());
-        }
-    }
-
-    public void chatHere() {
+    public void run() {
+        ui.welcome();
 
         Scanner sc = new Scanner(System.in);
-        String command = sc.next();
+        Parser parser = new Parser(tasks, ui, sc);
 
         while (true) {
-            switch (command) {
-                case "list":
-                    listTask();
-                    break;
-
-                case "bye":
-                    saveTasksToFile();
-                    goodbye();
-                    sc.close();
-                    return;
-
-                case "delete":
-                    try {
-                        int taskNum = sc.nextInt();
-                        deleteTask(taskNum);
-                    } catch (DinoException e) {
-                        System.out.println("Error: " + e.getMessage());
-                    }
-                    break;
-
-                case "todo":
-                    taskType = TaskType.TODO;
-                    handleTaskCreation(sc, taskType);
-                    break;
-
-                case "deadline":
-                    taskType = TaskType.DEADLINE;
-                    handleTaskCreation(sc, taskType);
-                    break;
-
-                case "event":
-                    taskType = TaskType.EVENT;
-                    handleTaskCreation(sc, taskType);
-                    break;
-
-                case "filter":
-                    printTasksForDate(sc);
-                    break;
-
-                case "mark":
-                    int taskNum = sc.nextInt();
-                    if (taskNum > taskList.size()) {
-                        System.out.println("Uh oh, we do not have a task assigned to that number.");
-                    } else {
-                        System.out.println("Good job on completing the task! I have checked it off the list.");
-                        Task completed = taskList.get(taskNum - 1);
-                        completed.markAsDone();
-                    }
-                    break;
-
-                case "unmark":
-                    int taskNumber = sc.nextInt();
-                    if (taskNumber > taskList.size()) {
-                        System.out.println("Uh oh, we do not have a task assigned to that number.");
-                    } else {
-                        System.out.println("Ah, I will mark it as undone. Remember to do it asap!");
-                        Task missing = taskList.get(taskNumber - 1);
-                        missing.markAsUndone();
-                    }
-                    break;
-
-                default:
-                    try {
-                        throw new DinoException("I don't understand ;;");
-                    } catch (DinoException e) {
-                        System.out.println("Error: " + e.getMessage());
-                    }
-                    break;
+            String command = sc.nextLine();
+            if (command.equals("bye")) {
+                storage.saveTasksToFile(tasks.getTaskList());
+                ui.goodbye();
+                sc.close();
+                break;
             }
-            command = sc.next();
+            parser.parseCommand(command);
         }
     }
 
-
     public static void main(String[] args) {
-        Dino mrDino = new Dino();
-        welcome();
-        mrDino.chatHere();
+        new Dino(FILE_PATH).run();
     }
 }
