+++ src/main/java/Ui.java
@@ -1,0 +1,9 @@
+public class Ui {
+    public void welcome() {
+        System.out.println("Hola! I'm Dino.\nWhat are you doing here?");
+    }
+
+    public void goodbye() {
+        System.out.println("Goodbye! It was nice meeting you.");
+    }
+}
